Use base64 extension in more packages
