+++ packages/playground/cli/vite.config.ts
@@ -7,6 +7,8 @@ import dts from 'vite-plugin-dts';
 import { viteTsConfigPaths } from '../../vite-extensions/vite-ts-config-paths';
 // eslint-disable-next-line @nx/enforce-module-boundaries
 import { getExternalModules } from '../../vite-extensions/vite-external-modules';
+// eslint-disable-next-line @nx/enforce-module-boundaries
+import viteGlobalExtensions from '../../vite-extensions/vite-global-extensions';
 
 /**
  * @TODO: Consider rsbuild for this:
@@ -86,6 +88,7 @@ const plugins = [
 			}
 		},
 	},
+	...viteGlobalExtensions,
 ] as PluginOption[];
 
 const external = [
